[Refactor]: 도메인 정의에 맞게 패키지 변경
Related to: #559
