+++ src/main/java/org/sopt/makers/internal/community/domain/anonymous/AnonymousNickname.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.internal.domain.community;
+package org.sopt.makers.internal.community.domain.anonymous;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -6,18 +6,17 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Entity
 @Getter
+@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class AnonymousNickname {
 
-	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "anonymous_nickname_id")
 	private Long id;
