+++ src/main/java/org/sopt/makers/internal/community/service/anonymous/AnonymousNicknameRetriever.java
@@ -2,7 +2,7 @@ package org.sopt.makers.internal.community.service.anonymous;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.makers.internal.community.repository.anonymous.AnonymousNicknameRepository;
-import org.sopt.makers.internal.domain.community.AnonymousNickname;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousNickname;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
