+++ src/main/java/org/sopt/makers/internal/community/service/anonymous/AnonymousPostProfileRetriever.java
@@ -2,7 +2,7 @@ package org.sopt.makers.internal.community.service.anonymous;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.makers.internal.community.repository.anonymous.AnonymousPostProfileRepository;
-import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousPostProfile;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
