+++ src/main/java/org/sopt/makers/internal/community/service/anonymous/AnonymousPostProfileService.java
@@ -1,11 +1,11 @@
 package org.sopt.makers.internal.community.service.anonymous;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.makers.internal.community.domain.AnonymousProfileImage;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousProfileImage;
 import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.sopt.makers.internal.domain.Member;
-import org.sopt.makers.internal.domain.community.AnonymousNickname;
-import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousNickname;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousPostProfile;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
