+++ src/main/java/org/sopt/makers/internal/community/service/anonymous/AnonymousProfileImageRetriever.java
@@ -1,7 +1,7 @@
 package org.sopt.makers.internal.community.service.anonymous;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.makers.internal.community.domain.AnonymousProfileImage;
+import org.sopt.makers.internal.community.domain.anonymous.AnonymousProfileImage;
 import org.sopt.makers.internal.community.repository.anonymous.AnonymousProfileImageRepository;
 import org.sopt.makers.internal.exception.BusinessLogicException;
 import org.springframework.stereotype.Component;
